% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcfraildv.control.R
\name{bcfraildv.control}
\alias{bcfraildv.control}
\title{Arguments for controlling bcfraildv fits.}
\usage{
bcfraildv.control(
  max.iter = 500,
  tol = 1e-04,
  eval.max = 500,
  iter.max = 500,
  trace = 0,
  abs.tol = 1e-20,
  rel.tol = 1e-10,
  x.tol = 1.5e-08,
  xf.tol = 2.2e-14,
  step.min = 1,
  step.max = 1
)
}
\arguments{
\item{max.iter}{Maximum number of iterations allowed. The default is 500.}

\item{tol}{A tolerance for convergence i.e the maximum differences of loglikelihood between succssive iterations.The default is 1e-04.}

\item{eval.max}{argument used to control \link{nlminb} fits used.}

\item{iter.max}{argument used to control \link{nlminb} fits used.}

\item{trace}{argument used to control \link{nlminb} fits used.}

\item{abs.tol}{argument used to control \link{nlminb} fits used.}

\item{rel.tol}{argument used to control \link{nlminb} fits used.}

\item{x.tol}{argument used to control \link{nlminb} fits used.}

\item{xf.tol}{argument used to control \link{nlminb} fits used.}

\item{step.min}{argument used to control \link{nlminb} fits used.}

\item{step.max}{argument used to control \link{nlminb} fits used.}
}
\value{
A list of control parameters.
}
\description{
This is used to set various numeric parameters controlling a bcfraildv model fit as a single list.
}
\seealso{
\code{\link{bcfraildv}}
}
