% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bcfraildv.R
\name{print.bcfraildv}
\alias{print.bcfraildv}
\title{Print bcfraildv}
\usage{
\method{print}{bcfraildv}(x, ...)
}
\arguments{
\item{x}{A class \code{bcfraildv} object.}

\item{...}{ignored}
}
\value{
An object of \code{print.bcfraildv}, with some more human-readable results from \code{bcfraildv} object.
}
\description{
Generics to print the S3 class bcfraildv.
}
\details{
Calls \code{print.bcfraildv()}.
}
\note{
The summary function is currently identical to the print function.
}
\examples{
set.seed(24)
simdata<-simbcfraildv(p.size=300, c.rate= c(0.3),fraildistrn=c("gamma"),frail.par=c(0.5,0.5,0.5),
bhaz.arg=list(distrn = c("weibull"),shape =c(5), scale = c(0.1)),
covar.arg=list(coefs=c(2),types = c("B"),size=c(1),prob=c(0.5)))
dataa<-simdata$data

fitbcfraildv=bcfraildv(Surv(time,censor)~ X1+frailty(PID) ,data=dataa,frail_distrn=c("gamma"))
fitbcfraildv
summary(fitbcfraildv)

}
\seealso{
\code{\link{bcfraildv}}
}
