% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bcfrailpar.R
\name{print.bcfrailpar}
\alias{print.bcfrailpar}
\title{Print bcfrailpar}
\usage{
\method{print}{bcfrailpar}(x, ...)
}
\arguments{
\item{x}{A class \code{bcfrailpar} object.}

\item{...}{ignored}
}
\value{
An object of \code{print.bcfrailpar}, with some more human-readable results from \code{bcfrailpar} object.
}
\description{
Generics to print the S3 class bcfrailpar.
}
\details{
Calls \code{print.bcfrailpar()}.
}
\note{
The summary function is currently identical to the print function.
}
\examples{
set.seed(4)
simdata<-simbcfraildv(psize=500, cenr= c(0),beta=c(-1),frailty=c("gamma"),
frailpar=c(0.4,0.4,0.5),bhaz=c("weibull"),
bhazpar=list(shape =c(0.9), scale = c(2)),
covartype= c("B"),covarpar=list(fargs=c(1),sargs=c(0.5)))
dataa<-simdata$data ## the simulated data set

#fit
parbcfit=bcfrailpar(Surv(time, censor) ~ X1+frailty(PID),data=dataa)
parbcfit

}
\seealso{
\code{\link{bcfrailpar}}
}
