% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bcgfrailev.R
\name{print.bcgfrailev}
\alias{print.bcgfrailev}
\title{Print bcgfrailev model}
\usage{
\method{print}{bcgfrailev}(x, ...)
}
\arguments{
\item{x}{A class \code{bcgfrailev} object.}

\item{...}{ignored}
}
\description{
Generics to print the S3 class bcgfrailev.
}
\details{
Calls \code{print.bcfrailph()}.
}
