% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bcfraildv.R
\name{summary.bcfraildv}
\alias{summary.bcfraildv}
\title{Print bcfraildv}
\usage{
\method{summary}{bcfraildv}(object, ...)
}
\arguments{
\item{object}{A class \code{bcfraildv} object.}

\item{...}{ignored}
}
\value{
An object of \code{print.bcfraildv}, with some more human-readable results from \code{bcfraildv} object.
}
\description{
Generics to print the S3 class bcfraildv.
}
\details{
Calls \code{print.bcfraildv()}.
}
\note{
The summary function is currently identical to the print function.
}
\examples{
set.seed(4)
simdata<-simbcfraildv(psize=300, cenr= c(0.3),beta=c(2),frailty=c("gamma"),
frailpar=c(0.5,0.5,0.5),bhaz=c("weibull"),
bhazpar=list(shape =c(5), scale = c(0.1)),
covartype= c("B"),covarpar=list(fargs=c(1),sargs=c(0.5)))
dataa<-simdata$data

fitbcfrail=bcfraildv(Surv(time,censor)~ X1+frailty(PID) ,data=dataa)
fitbcfrail
summary(fitbcfrail)

}
\seealso{
\code{\link{bcfraildv}}
}
