% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bcfrailpar.R
\name{summary.bcfrailpar}
\alias{summary.bcfrailpar}
\title{Print bcfrailpar}
\usage{
\method{summary}{bcfrailpar}(object, ...)
}
\arguments{
\item{object}{A class \code{bcfrailpar} object.}

\item{...}{ignored}
}
\value{
An object of \code{summary.bcfrailpar}, with some more human-readable results from \code{bcfrailpar} object.
}
\description{
Generics to print the S3 class bcfrailpar.
}
\details{
Calls \code{print.bcfrailpar()}.
}
\note{
The summary function is currently identical to the print function.
}
\examples{
set.seed(40)
simdata<-simbcfraildv(psize=500, cenr= c(0.3),beta=c(-1),frailty=c("gamma"),
frailpar=c(0.4,0.4,0.5),bhaz=c("gompertz"),
bhazpar=list(shape =c(0.09), scale = c(0.2)),
covartype= c("B"),covarpar=list(fargs=c(1),sargs=c(0.5)))
dataa<-simdata$data

fitbcfrail=bcfrailpar(Surv(time,censor)~ X1+frailty(PID) ,
data=dataa,haz="gompertz")
fitbcfrail
summary(fitbcfrail)

}
\seealso{
\code{\link{bcfrailpar}}
}
