% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simbcfraildv.R
\name{simbcfraildv}
\alias{simbcfraildv}
\title{Simulate data from bivariate correlated frailty models.}
\usage{
simbcfraildv(
  p.size,
  c.rate = c(0),
  fraildistrn,
  frail.par = c(0.5, 0.5, 0.25),
  bhaz.arg = list(distrn = c("weibull"), shape = c(2.5), scale = c(0.01), rate =
    c(0.5)),
  covar.arg = list(coefs = c(0.5, 0), types = c("BU", "BB", "UU", "B", "U"), size =
    c(1, 1), prob = c(0.5, 0.6), min = c(0, 0), max = c(1, 3))
)
}
\arguments{
\item{p.size}{pair size.}

\item{c.rate}{censored rate. The default is zero..}

\item{fraildistrn}{A type of frailty distribution to be used. Either gamma or log-normal.}

\item{frail.par}{vector of frailty parameters, variance and correlation respectively. The default is c(0.5,0.5,0.25) meaning frailty variance of subject1 is 0.5, frailty variance of subject2 is 0.5, and correlation 0.25.}

\item{bhaz.arg}{is a \code{\link{list}} i.e,\code{list(distrn = c("weibull"),shape =c(2.5), scale = c(0.01), rate = c(0.5))}.\code{distrn} is the type of baseline hazard to be used. weibull, gompertz and exponential is possible. \code{shape}, \code{scale} and \code{rate} are the parameters of the coresponding baseline hazard parameters. rate needs to be specified if the baseline is exponential.}

\item{covar.arg}{is a \code{\link{list}} i.e,\code{coefs} is covariate coefficients, \code{types} is the type of covariate to be used. One of the following can be specified \code{c("BU","BB","UU","B","U")}. Use \code{"BU"} if one binomial and one uniform covariates are desired.\code{"BB"} is for two binomial covariates. \code{"B"} is for binomial and \code{"U"} is for uniform.\code{size} and \code{prob} needs to be specified if binomial covariate(s) are used and if uniform covariate(s) are used, then \code{min} and \code{max} should be specified.}
}
\value{
An object of class \code{simbcfraildv} that contain the following:

\itemize{

\item{\code{data}}  {A data frame i.e, the simulated data set. IID is individual Id, PID is pair ID, time is the simulated survival time, censor is censoring indicator and X1 or X2 denote the simulated covariate.}

\item{\code{numberofpair}}  {The specified number of pairs.}

\item{\code{censoredrate} } {The specified censored rate.}

\item{\code{fraildist} } {The specified frailty distribution.}

\item{\code{frailpar}} {The specified frailty parameters.}
}
}
\description{
Simulate data from bivariate correlated gamma and log-normal frailty models with observed covariates. It allows inclussion of one or two covariates. In addition, frailty variances of the two artificial subjects can be equal or different.
}
\examples{
simdata<-simbcfraildv(p.size=1000, c.rate= c(0.2),fraildistrn=c("gamma"),frail.par=c(0.5,0.5,0.5),
bhaz.arg=list(distrn = c("gompertz"),shape =c(3), scale = c(0.1)),
covar.arg=list(coefs=c(1),types = c("U"),min=0,max=1))
simdata

\donttest{
#Let us simulate a data set with the following parameters
#weibull baseline hazard with parameters shape= 2.5 and scale=0.01.
#a dataset with 1000 pairs. Frailty distribution is gamma
# and the frailty parameters are taken to
#be variance1=0.6,variance2=0.4 and correlation =0.2.
#One binomial covariate i.e,
#(Binomial (1,0.5)) with regression coefficient 0.5.
#Each observed covariate for the two individuals in a
#pair is taken to be independent and 20 percent
#observations are censored.

#simulate the data set

set.seed(1)
simdata<-simbcfraildv(p.size=1000, c.rate= c(0.2),
fraildistrn=c("gamma"),frail.par=c(0.6,0.4,0.2),
bhaz.arg=list(distrn = c("weibull"),shape =c(2.5), scale = c(0.01)),
covar.arg=list(coefs=c(0.5),types = c("B"),size=1,prob=0.5))

#to extract the simulated data set
dataa<-simdata$data ## the simulated data set
dataa[1:4,] # the first four rows looks like

#  IID PID     time censor X1
#1   1   1 2.704927      1  1
#2   2   1 5.418071      0  1
#3   3   2 4.602736      1  0
#4   4   2 6.205303      1  0

# IID is individual indicator
# PID is pair indicator
# time is the simulated survival time
# censor is the simulated censoring indicator
# X1 is the simulated covariate

# if log-normal frailty is desired

simdata<-simbcfraildv(p.size=1000, c.rate= c(0.2),
fraildistrn=c("lognormal"),frail.par=c(0.6,0.4,0.2),
bhaz.arg=list(distrn = c("weibull"),shape =c(2.5), scale = c(0.01)),
covar.arg=list(coefs=c(0.5),types = c("B"),size=1,prob=0.5))

dataa<-simdata$data # the simulated data set

# if log-normal frailty with two covariates
#i.e., binomial (Binomial (1,0.5)) and uniform U[0,1] is desired

simdata<-simbcfraildv(p.size=1000, c.rate= c(0.2),
fraildistrn=c("lognormal"),frail.par=c(0.6,0.4,0.2),
bhaz.arg=list(distrn = c("weibull"),shape =c(2.5), scale = c(0.01)),
covar.arg=list(coefs=c(0.5),types = c("BU"),
size=1,prob=0.5,min=c(0),max=c(1)))

dataa<-simdata$data ## the simulated data set

# Not run
# if p.size, pair size missed
simdata<-simbcfraildv( c.rate= c(0.2),fraildistrn=c("gamma"),
frail.par=c(0.6,0.4,0.2),
covar.arg=list(coefs=c(0.5),types = c("B"),size=1,prob=0.5))

# if frailty distribution other than gamma and lognormal specified

simdata<-simbcfraildv(p.size=100, c.rate= c(0.2),fraildistrn=c("exp"),frail.par=c(0.6,0.4,0.6),
covar.arg=list(coefs=c(0.5),types = c("B"),size=1,prob=0.5))
# End Not run
}

}
\seealso{
\code{\link{bcfraildv}}
}
